Fix: value being int

+++ moves/Defend.py
@@ -10,7 +10,7 @@ class Defend(Move):
     
     def effect(self, owner: Creature):
         if (self.is_take):
-            value = self.value * self.rate
+            value = int(self.value * self.rate)
             owner.defense += value
     
     def __str__(self) -> str:
